break App.tsx into smaller files
