+++ termpair/frontend_src/src/App.tsx
@@ -1,9 +1,5 @@
 import React, { useEffect, useState, useLayoutEffect, useRef } from "react";
 import "xterm/css/xterm.css";
-import logo from "./logo.png"; // logomakr.com/4N54oK
-import { TERMPAIR_VERSION } from "./constants";
-// import { CogIcon } from "@heroicons/react/solid";
-import { DuplicateIcon } from "@heroicons/react/solid";
 import { Terminal as Xterm, IDisposable } from "xterm";
 import moment from "moment";
 import {
@@ -14,7 +10,6 @@ import {
 } from "./encryption";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-// import { atom, useRecoilState } from "recoil";
 import { debounce } from "debounce";
 import {
   newBrowserConnected,
@@ -22,79 +17,9 @@ import {
   requestTerminalDimensions,
   sendCommandToTerminal,
 } from "./events";
-import { CopyToClipboard } from "react-copy-to-clipboard";
-
-const githubLogo = (
-  <svg
-    width="24"
-    height="24"
-    fill="currentColor"
-    className="text-gray-300 mr-3 "
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M12 2C6.477 2 2 6.463 2 11.97c0 4.404 2.865 8.14 6.839 9.458.5.092.682-.216.682-.48 0-.236-.008-.864-.013-1.695-2.782.602-3.369-1.337-3.369-1.337-.454-1.151-1.11-1.458-1.11-1.458-.908-.618.069-.606.069-.606 1.003.07 1.531 1.027 1.531 1.027.892 1.524 2.341 1.084 2.91.828.092-.643.35-1.083.636-1.332-2.22-.251-4.555-1.107-4.555-4.927 0-1.088.39-1.979 1.029-2.675-.103-.252-.446-1.266.098-2.638 0 0 .84-.268 2.75 1.022A9.606 9.606 0 0112 6.82c.85.004 1.705.114 2.504.336 1.909-1.29 2.747-1.022 2.747-1.022.546 1.372.202 2.386.1 2.638.64.696 1.028 1.587 1.028 2.675 0 3.83-2.339 4.673-4.566 4.92.359.307.678.915.678 1.846 0 1.332-.012 2.407-.012 2.734 0 .267.18.577.688.48C19.137 20.107 22 16.373 22 11.969 22 6.463 17.522 2 12 2z"
-    ></path>
-  </svg>
-);
-
-// const showSettings = atom({
-//   key: "showSettings",
-//   default: false,
-// });
-
-// function Settings(props: any) {
-//   const [showSetting, setShowSettings] = useRecoilState(showSettings);
-//   if (!showSetting) {
-//     return null;
-//   }
-//   return (
-//     <div
-//       className="w-full h-full bg-gray-900 absolute bg-opacity-90  text-black"
-//       style={{ zIndex: 2000 }}
-//     >
-//       <div className="w-11/12 h-5/6 m-10 p-5 bg-gray-400 flex items-center flex-col">
-//         <div className="text-xl mb-10">TermPair Settings</div>
-//         <div className="flex-grow">Body</div>
-//         <div>
-//           <button
-//             className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full"
-//             onClick={() => setShowSettings(false)}
-//           >
-//             Close
-//           </button>
-//         </div>
-//       </div>
-//     </div>
-//   );
-// }
-
-function TopBar(props: any) {
-  return (
-    <div className="flex bg-black h-10 items-center justify-between">
-      <div className="h-full">
-        <a href={window.location.pathname}>
-          <img className="h-full" src={logo} alt="logo" />
-        </a>
-      </div>
-      <div className="flex">
-        <span className="text-gray-300 mx-3">v{TERMPAIR_VERSION}</span>
-        <a href="https://github.com/cs01/termpair">{githubLogo}</a>
-      </div>
-      {/* <div className="text-white m-5">
-        <div className="my-auto">
-          <button
-            className="my-auto"
-            onClick={() => setShowSettings(!showSetting)}
-          >
-            <CogIcon className="h-6 w-6 text-white" />
-          </button>
-        </div>
-      </div> */}
-    </div>
-  );
-}
+import { LandingPageContent } from "./LandingPageContent";
+import { AesKeysRef } from "./types";
+import { TopBar } from "./TopBar";
 
 function BottomBar(props: {
   status: Status;
@@ -194,271 +119,6 @@ class ErrorBoundary extends React.Component<any, any> {
 const cannotTypeMsg =
   "Terminal was shared in read only mode. Unable to send data to terminal's input.";
 
-const host = `${window.location.protocol}//${window.location.hostname}${window.location.pathname}`;
-let port = window.location.port;
-if (!window.location.port) {
-  if (window.location.protocol === "https:") {
-    port = "443";
-  } else {
-    port = "80";
-  }
-}
-const termpairShareCommand = `termpair share --host "${host}" --port ${port}`;
-const pipxTermpairShareCommand = `pipx run ${termpairShareCommand}`;
-
-function CopyCommand(props: { command: string }) {
-  const [clicked, setClicked] = useState(false);
-  const [hovering, setHovering] = useState(false);
-  return (
-    <div className="flex">
-      <code
-        className={`${
-          hovering || clicked ? "bg-yellow-200" : "bg-gray-300"
-        } text-black px-2 py-1 m-2`}
-      >
-        {props.command}
-      </code>
-      <CopyToClipboard text={props.command}>
-        <button
-          className="px-2"
-          title="Copy command to clipboard"
-          onMouseEnter={() => setHovering(true)}
-          onMouseLeave={() => setHovering(false)}
-          onClick={() => {
-            setClicked(true);
-            setTimeout(() => setClicked(false), 1500);
-          }}
-        >
-          <DuplicateIcon className="h-6 w-6 text-white" />
-        </button>
-      </CopyToClipboard>
-      <span className="py-1 m-2">{clicked ? "Copied!" : ""}</span>
-    </div>
-  );
-}
-
-function LandingPageContent(props: {
-  isStaticallyHosted: Nullable<boolean>;
-  setCustomTermpairServer: (customServer: string) => void;
-  setTerminalId: (newTerminalId: string) => void;
-  aesKeys: React.MutableRefObject<AesKeysRef>;
-}) {
-  const [terminalIdInput, setTerminalIdInput] = React.useState("");
-  const [customHostInput, setCustomHostInput] = React.useState("");
-  const [bootstrapAesKeyB64Input, setBootstrapAesKeyB64Input] =
-    React.useState("");
-
-  const submitForm = async () => {
-    if (!terminalIdInput) {
-      toast.dark("Terminal ID cannot be empty");
-      return;
-    }
-    if (!bootstrapAesKeyB64Input) {
-      toast.dark("Secret key cannot be empty");
-      return;
-    }
-    if (!props.isStaticallyHosted) {
-      props.setTerminalId(terminalIdInput);
-    }
-    if (!customHostInput) {
-      toast.dark("Host name cannot be empty");
-      return;
-    }
-    try {
-      new URL(customHostInput);
-    } catch (e) {
-      toast.dark(`${customHostInput} is not a valid url`);
-      return;
-    }
-    try {
-      const bootstrapKey = await getAESKey(
-        Buffer.from(bootstrapAesKeyB64Input, "base64"),
-        ["decrypt"]
-      );
-      props.aesKeys.current.bootstrap = bootstrapKey;
-    } catch (e) {
-      toast.dark(`Secret encryption key is not valid`);
-      return;
-    }
-    props.setCustomTermpairServer(customHostInput);
-    props.setTerminalId(terminalIdInput);
-  };
-  const inputClass = "text-black px-2 py-3 m-2 w-full font-mono";
-
-  const terminalIdInputEl = (
-    <div
-      className="flex items-center"
-      title="The unique Terminal ID the broadcasting terminal was provided when the sharing session began."
-    >
-      <span className="py-2 m-2 whitespace-nowrap text-xl">Terminal ID</span>
-      <input
-        name="terminalIdInput"
-        type="text"
-        className={inputClass}
-        onChange={(event) => {
-          setTerminalIdInput(event.target.value);
-        }}
-        value={terminalIdInput}
-        placeholder="abcdef123456789abcded123456789"
-      />
-    </div>
-  );
-  const bootstrapCryptoKeyInputEl = (
-    <div className="flex items-center" title="Base 64 encoded AES key">
-      <span className="py-2 m-2 whitespace-nowrap text-xl">
-        Secret encryption key
-      </span>
-      <input
-        name="bootstrapAesKeyB64Input"
-        placeholder="123456789abcded123456789"
-        type="text"
-        className={inputClass}
-        onChange={(event) => {
-          setBootstrapAesKeyB64Input(event.target.value);
-        }}
-        value={bootstrapAesKeyB64Input}
-      />
-    </div>
-  );
-  const terminalServerUrlEl = (
-    <div
-      className="flex items-center"
-      title="The URL of an actual TermPair server that the terminal is broadcasting through."
-    >
-      <span className="py-2 m-2 whitespace-nowrap text-xl">
-        TermPair Server URL
-      </span>
-      <input
-        name="customHostInput"
-        type="text"
-        className={inputClass}
-        placeholder="http://localhost:8000"
-        onChange={(event) => {
-          setCustomHostInput(event.target.value);
-        }}
-        value={customHostInput}
-      />
-    </div>
-  );
-
-  const canConnect = props.isStaticallyHosted
-    ? terminalIdInput.length !== 0
-    : terminalIdInput.length !== 0 && customHostInput.length !== 0;
-
-  const connectButton = (
-    <div className="flex justify-end">
-      <button
-        type="submit"
-        title="Connect to the specified Terminal"
-        className={`bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-full ${
-          canConnect ? "" : "cursor-not-allowed"
-        }`}
-      >
-        Connect
-      </button>
-    </div>
-  );
-  const connectForm = (
-    <form
-      onSubmit={(e) => {
-        e.preventDefault();
-        submitForm();
-      }}
-    >
-      {terminalIdInputEl}
-      {bootstrapCryptoKeyInputEl}
-      {props.isStaticallyHosted ? terminalServerUrlEl : null}
-      {connectButton}
-    </form>
-  );
-  const staticLandingContent = (
-    <div className="py-2">
-      <div className="text-2xl py-2">This page is statically hosted</div>
-      <div>
-        This is a static page serving the TermPair JavaScript app. It is
-        optional to use a statically served TermPair webapp, but it facilitates
-        easily building and self-serving to be certain the JavaScript app has
-        not been tampered with by an untrusted server.
-      </div>
-      <div className="mt-5">
-        Connect to a broadcasting terminal by entering the fields below and
-        clicking Connect.
-      </div>
-      {connectForm}
-    </div>
-  );
-
-  const regularServerContent = (
-    <>
-      <div className="py-2">
-        <div className="text-xl  py-2">Quick Start</div>
-        <div>
-          If you have TermPair installed, share a terminal with this host:
-        </div>
-        <CopyCommand command={termpairShareCommand} />
-        <div>Or if you have pipx, you can run TermPair via pipx:</div>
-        <CopyCommand command={pipxTermpairShareCommand} />
-      </div>
-      <div className="py-2">
-        <div className="text-xl  py-2">Install TermPair</div>
-        <div>Install with pipx</div>
-        <CopyCommand command="pipx install termpair" />
-        <div>Or install with pip</div>
-        <CopyCommand command="pip install termpair --user" />
-      </div>
-      <div className="py-2">
-        <div className="text-xl  py-2">Connecting to a Terminal?</div>
-        If a terminal is already broadcasting and you'd like to connect to it,
-        you don't need to install or run anything. Just fill out the form below
-        and click Connect.
-        {connectForm}
-      </div>
-    </>
-  );
-
-  const termpairDemoContent = (
-    <div className="py-2">
-      <div className="text-xl  py-2">TermPair Demo</div>
-      <div>
-        <img
-          alt="Screencast of TermPair"
-          src="https://raw.githubusercontent.com/cs01/termpair/master/termpair_browser.gif"
-        />
-      </div>
-    </div>
-  );
-
-  return (
-    <div className="flex justify-center">
-      <div className="text-gray-200 max-w-3xl">
-        <div className="py-2">
-          <div className="text-3xl ">Welcome to TermPair!</div>
-          Easily share terminals with end-to-end encryption 🔒. Terminal data is
-          always encrypted before being routed through the server.{" "}
-          <a href="https://github.com/cs01/termpair">Learn more.</a>
-        </div>
-        {props.isStaticallyHosted === null
-          ? null
-          : props.isStaticallyHosted === true
-          ? staticLandingContent
-          : regularServerContent}
-
-        <div className="py-2">
-          <div className="text-2xl py-2">Troubleshooting</div>
-          <div className="text-xl ">
-            Initial connection fails or is rejected
-          </div>
-          <div>
-            Ensure you are using a TermPair client compatible with{" "}
-            <span className="font-bold">v{TERMPAIR_VERSION}</span> (the version
-            of this webpage)
-          </div>
-        </div>
-        {termpairDemoContent}
-      </div>
-    </div>
-  );
-}
 type Status =
   | null
   | "Connecting..."
@@ -590,13 +250,6 @@ function handleStatusChange(
   return status as never;
 }
 
-type AesKeysRef = {
-  bootstrap: Nullable<CryptoKey>;
-  browser: Nullable<CryptoKey>;
-  unix: Nullable<CryptoKey>;
-  ivCount: Nullable<number>;
-  maxIvCount: Nullable<number>;
-};
 function App() {
   const [isStaticallyHosted, setIsStaticallyHosted] =
     useState<Nullable<boolean>>(null);
