+++ termpair/frontend_src/src/types.tsx
@@ -1,0 +1,7 @@
+export type AesKeysRef = {
+  bootstrap: Nullable<CryptoKey>;
+  browser: Nullable<CryptoKey>;
+  unix: Nullable<CryptoKey>;
+  ivCount: Nullable<number>;
+  maxIvCount: Nullable<number>;
+};
